+++ spark/src/main/scala/geotrellis/spark/io/cog/COGLayer.scala
@@ -41,9 +41,9 @@ object COGLayer {
     *
     * @param rdd             Layer layer, at highest resolution
     * @param baseZoom        Zoom level of the base layer, assumes [[ZoomedLayoutScheme]]
-    * @param compression     Compression method for GeoTiff tiles
     * @param minZoom         Zoom level at which to stop the pyramiding.
-    * @param options         [[COGLayerWriter.Options]] that contains information on the maxTileSize of the layer.
+    * @param options         [[COGLayerWriter.Options]] that contains information on the maxTileSize,
+    *                        resampleMethod, and compression of the written layer.
     */
   def fromLayerRDD[
     K: SpatialComponent: Ordering: JsonFormat: ClassTag,
@@ -51,9 +51,8 @@ object COGLayer {
   ](
      rdd: RDD[(K, V)] with Metadata[TileLayerMetadata[K]],
      baseZoom: Int,
-     compression: Compression = Deflate,
      minZoom: Option[Int] = None,
-     options: COGLayerWriter.Options
+     options: COGLayerWriter.Options = COGLayerWriter.Options.DEFAULT
    ): COGLayer[K, V] = {
     // TODO: Clean up conditional checks, figure out how to bake into type system, or report errors better.
 
@@ -66,6 +65,8 @@ object COGLayer {
         sys.error("Cannot create Pyramided COG layer for tile sizes that are not power-of-two.")
       }
     }
+
+    val compression = options.compression
 
     val layoutScheme =
       ZoomedLayoutScheme(rdd.metadata.crs, rdd.metadata.layout.tileCols)