+++ spark/src/main/scala/geotrellis/spark/io/cog/COGLayerWriter.scala
@@ -60,7 +60,7 @@ trait COGLayerWriter extends LazyLogging with Serializable {
      tiles: RDD[(K, V)] with Metadata[TileLayerMetadata[K]],
      tileZoom: Int,
      keyIndexMethod: KeyIndexMethod[K]
-   ): Unit = write[K, V](layerName, tiles, tileZoom, keyIndexMethod, NoCompression, None, Options.DEFAULT)
+   ): Unit = write[K, V](layerName, tiles, tileZoom, keyIndexMethod, None, Options.DEFAULT)
 
   def write[
     K: SpatialComponent: Ordering: JsonFormat: ClassTag,
@@ -71,7 +71,7 @@ trait COGLayerWriter extends LazyLogging with Serializable {
      tileZoom: Int,
      keyIndexMethod: KeyIndexMethod[K],
      options: Options
-   ): Unit = write[K, V](layerName, tiles, tileZoom, keyIndexMethod, NoCompression, None, options)
+   ): Unit = write[K, V](layerName, tiles, tileZoom, keyIndexMethod, None, options)
 
   def write[
     K: SpatialComponent: Ordering: JsonFormat: ClassTag,
@@ -81,13 +81,12 @@ trait COGLayerWriter extends LazyLogging with Serializable {
     tiles: RDD[(K, V)] with Metadata[TileLayerMetadata[K]],
     tileZoom: Int,
     keyIndexMethod: KeyIndexMethod[K],
-    compression: Compression,
     mergeFunc: Option[(GeoTiff[V], GeoTiff[V]) => GeoTiff[V]],
     options: Options
   ): Unit =
     tiles.metadata.bounds match {
       case keyBounds: KeyBounds[K] =>
-        val cogLayer = COGLayer.fromLayerRDD(tiles, tileZoom, compression = compression, options = options)
+        val cogLayer = COGLayer.fromLayerRDD(tiles, tileZoom, options = options)
         // println(cogLayer.metadata.toJson.prettyPrint)
         val keyIndexes: Map[ZoomRange, KeyIndex[K]] =
           cogLayer.metadata.zoomRangeInfos.
@@ -106,7 +105,7 @@ trait COGLayerWriter extends LazyLogging with Serializable {
      tiles: RDD[(K, V)] with Metadata[TileLayerMetadata[K]],
      tileZoom: Int,
      keyIndex: KeyIndex[K]
-   ): Unit = write[K, V](layerName, tiles, tileZoom, keyIndex, NoCompression, None, Options.DEFAULT)
+   ): Unit = write[K, V](layerName, tiles, tileZoom, keyIndex, None, Options.DEFAULT)
 
   def write[
     K: SpatialComponent: Ordering: JsonFormat: ClassTag,
@@ -117,7 +116,7 @@ trait COGLayerWriter extends LazyLogging with Serializable {
      tileZoom: Int,
      keyIndex: KeyIndex[K],
      options: Options
-   ): Unit = write[K, V](layerName, tiles, tileZoom, keyIndex, NoCompression, None, options)
+   ): Unit = write[K, V](layerName, tiles, tileZoom, keyIndex, None, options)
 
   def write[
     K: SpatialComponent: Ordering: JsonFormat: ClassTag,
@@ -127,13 +126,12 @@ trait COGLayerWriter extends LazyLogging with Serializable {
      tiles: RDD[(K, V)] with Metadata[TileLayerMetadata[K]],
      tileZoom: Int,
      keyIndex: KeyIndex[K],
-     compression: Compression,
      mergeFunc: Option[(GeoTiff[V], GeoTiff[V]) => GeoTiff[V]],
      options: Options
    ): Unit =
     tiles.metadata.bounds match {
       case keyBounds: KeyBounds[K] =>
-        val cogLayer = COGLayer.fromLayerRDD(tiles, tileZoom, compression = compression, options = options)
+        val cogLayer = COGLayer.fromLayerRDD(tiles, tileZoom, options = options)
         val keyIndexes: Map[ZoomRange, KeyIndex[K]] =
           cogLayer.metadata.zoomRangeInfos.
             map { case (zr, _) => zr -> keyIndex }.
@@ -169,10 +167,9 @@ trait COGLayerWriter extends LazyLogging with Serializable {
     layerName: String,
     tiles: RDD[(K, V)] with Metadata[TileLayerMetadata[K]],
     tileZoom: Int,
-    compression: Compression = NoCompression,
     options: Options = Options.DEFAULT
   ): Unit =
-    if(tiles.metadata.bounds.nonEmpty) update[K, V](layerName, tiles, tileZoom, compression, None, options)
+    if(tiles.metadata.bounds.nonEmpty) update[K, V](layerName, tiles, tileZoom, None, options)
     else logger.info("Skipping layer update with empty bounds rdd.")
 
   def update[
@@ -182,7 +179,6 @@ trait COGLayerWriter extends LazyLogging with Serializable {
      layerName: String,
      tiles: RDD[(K, V)] with Metadata[TileLayerMetadata[K]],
      tileZoom: Int,
-     compression: Compression = NoCompression,
      mergeFunc: Option[(GeoTiff[V], GeoTiff[V]) => GeoTiff[V]],
      options: Options = Options.DEFAULT
    ): Unit = {
@@ -202,7 +198,7 @@ trait COGLayerWriter extends LazyLogging with Serializable {
         if(!indexKeyBounds.contains(keyBounds) && !metadata.keyBoundsForZoom(tileZoom).contains(keyBounds))
           throw new LayerOutOfKeyBoundsError(LayerId(layerName, tileZoom), indexKeyBounds)
 
-        val cogLayer = COGLayer.fromLayerRDD(tiles, tileZoom, compression = compression, options = options)
+        val cogLayer = COGLayer.fromLayerRDD(tiles, tileZoom, options = options)
         val ucogLayer = cogLayer.copy(metadata = cogLayer.metadata.combine(metadata))
 
         writeCOGLayer(layerName, ucogLayer, keyIndexes, mergeFunc)
@@ -216,7 +212,8 @@ object COGLayerWriter {
 
   case class Options(
     maxTileSize: Int = DefaultMaxTileSize,
-    resampleMethod: ResampleMethod = NearestNeighbor
+    resampleMethod: ResampleMethod = NearestNeighbor,
+    compression: Compression = NoCompression
   )
 
   object Options {
@@ -224,6 +221,7 @@ object COGLayerWriter {
 
     implicit def maxTileSizeToOptions(maxTileSize: Int): Options = Options(maxTileSize = maxTileSize)
     implicit def resampleMethodToOptions(resampleMethod: ResampleMethod): Options = Options(resampleMethod = resampleMethod)
+    implicit def compressionToOption(compression: Compression): Options = Options(compression = compression)
   }
 
   private val DefaultMaxTileSize = 4096